renaming templates to modelos
