+++ apps/application/convex/agent/tools/modelos/searchTemplatesTool.ts
