+++ apps/application/src/pages/ModelsPage.tsx
@@ -3,7 +3,7 @@ import {
   TemplateTable, 
   TemplatePreviewDialog, 
   TemplateSearchBar 
-} from "@/components/Templates";
+} from "@/components/Modelos";
 import { useQuery, useMutation } from "convex/react";
 import { api } from "../../convex/_generated/api";
 import type { Id } from "../../convex/_generated/dataModel";
